+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Add support for subclassing using `Class.new` to `Lint/InheritException`. ([@houli][])
 * [#6779](https://github.com/rubocop-hq/rubocop/issues/6779): Add new cop `Style/NegativeUnless` that checks for unless with negative condition. ([@tejasbubane][])
+* Add `IgnoreMethodPatterns` config option to `Style/MethodCallWithArgsParentheses`. ([@tejasbubane][])
 
 ### Bug fixes